add bi-direction file translator
