+++ hedera-node/hedera-file-service-impl/src/main/java/com/hedera/node/app/service/file/impl/codec/FileServiceStateTranslator.java
@@ -1,0 +1,111 @@
+/*
+ * Copyright (C) 2023 Hedera Hashgraph, LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.hedera.node.app.service.file.impl.codec;
+
+import static java.util.Objects.requireNonNull;
+
+import com.hedera.hapi.node.base.FileID;
+import com.hedera.hapi.node.base.Key;
+import com.hedera.hapi.node.state.file.File;
+import com.hedera.node.app.service.file.impl.ReadableFileStoreImpl;
+import com.hedera.node.app.service.mono.files.HFileMeta;
+import com.hedera.node.app.service.mono.pbj.PbjConverter;
+import com.hedera.pbj.runtime.io.buffer.Bytes;
+import edu.umd.cs.findbugs.annotations.NonNull;
+import org.apache.commons.codec.DecoderException;
+
+/**
+ * The class is used to convert a {@link com.hedera.node.app.service.mono.files.HFileMeta} content and metadata  to a {@link File} and vice versa during the migration process
+ */
+public class FileServiceStateTranslator {
+
+    /**
+     * The method converts a {@link com.hedera.node.app.service.mono.files.HFileMeta} content and metadata  to a {@link File}
+     * @param fileID old protobuf fileID that will be used to search in old state
+     * @param hederaFs old state
+     * @return new protobuf File
+     */
+    @NonNull
+    public static File stateToPbj(
+            @NonNull com.hederahashgraph.api.proto.java.FileID fileID,
+            @NonNull com.hedera.node.app.service.mono.files.HederaFs hederaFs) {
+        requireNonNull(fileID);
+        return stateToPbj(hederaFs.cat(fileID), hederaFs.getattr(fileID), fileID);
+    }
+
+    /**
+     * The method converts a {@link com.hedera.node.app.service.mono.files.HFileMeta} content and metadata  to a {@link File}
+     * @param data content of the file in old state
+     * @param metadata metadata of the file in old state
+     * @param fileID old protobuf fileID that will be used in order to initialize the new protobuf file
+     * @return new protobuf File
+     */
+    @NonNull
+    public static File stateToPbj(
+            @NonNull final byte[] data,
+            @NonNull final com.hedera.node.app.service.mono.files.HFileMeta metadata,
+            @NonNull final com.hederahashgraph.api.proto.java.FileID fileID) {
+        requireNonNull(data);
+        requireNonNull(metadata);
+        requireNonNull(fileID);
+        final var fileBuilder = new File.Builder();
+        fileBuilder.fileNumber(fileID.getFileNum());
+        fileBuilder.expirationTime(metadata.getExpiry());
+        fileBuilder.keys(PbjConverter.asPbjKey(metadata.getWacl()).keyList());
+        fileBuilder.contents(Bytes.wrap(data));
+        fileBuilder.memo(metadata.getMemo());
+        fileBuilder.deleted(metadata.isDeleted());
+
+        return fileBuilder.build();
+    }
+
+    /**
+     * The method converts a {@link File} to a {@link com.hedera.node.app.service.mono.files.HFileMeta} content and metadata
+     * @param fileID new protobuf fileID that will be used to search in file store
+     * @param readableFileStore file store that will be used to search for the file
+     * @return File and metadata pair object
+     */
+    @NonNull
+    public static FileMetadataAndContent pbjToState(
+            @NonNull FileID fileID, @NonNull ReadableFileStoreImpl readableFileStore) throws DecoderException {
+        requireNonNull(fileID);
+        requireNonNull(readableFileStore);
+        final var optionalFile = readableFileStore.getFileLeaf(fileID);
+        return pbjToState(optionalFile.orElseThrow(() -> new IllegalArgumentException("File not found")));
+    }
+
+    /**
+     * The method converts a {@link File} to a {@link com.hedera.node.app.service.mono.files.HFileMeta} content and metadata
+     * @param file new protobuf file that will be used
+     * @return File and metadata pair object
+     */
+    @NonNull
+    public static FileMetadataAndContent pbjToState(@NonNull File file) throws DecoderException {
+        requireNonNull(file);
+        var keys = (file.hasKeys())
+                ? com.hedera.node.app.service.mono.legacy.core.jproto.JKey.convertKey(
+                        Key.newBuilder().keyList(file.keys()).build(), 1)
+                : null;
+        com.hedera.node.app.service.mono.files.HFileMeta hFileMeta =
+                new HFileMeta(file.deleted(), keys, file.expirationTime(), file.memo());
+        return new FileMetadataAndContent(file.contents().toByteArray(), hFileMeta);
+    }
+
+    @SuppressWarnings("java:S6218")
+    public record FileMetadataAndContent(
+            @NonNull byte[] data, @NonNull com.hedera.node.app.service.mono.files.HFileMeta metadata) {}
+}
