+++ hedera-node/hedera-file-service-impl/src/main/java/module-info.java
@@ -14,6 +14,7 @@ module com.hedera.node.app.service.file.impl {
     requires com.swirlds.config;
     requires org.apache.commons.lang3;
     requires org.apache.logging.log4j;
+    requires org.apache.commons.codec;
 
     provides com.hedera.node.app.service.file.FileService with
             FileServiceImpl;
