+++ hedera-node/hedera-file-service-impl/src/test/java/com/hedera/node/app/service/file/impl/test/codec/FileServiceStateTranslatorTest.java
@@ -1,0 +1,129 @@
+/*
+ * Copyright (C) 2023 Hedera Hashgraph, LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.hedera.node.app.service.file.impl.test.codec;
+
+import static org.junit.jupiter.api.Assertions.assertArrayEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.mockito.BDDMockito.given;
+
+import com.hedera.hapi.node.base.Key;
+import com.hedera.hapi.node.state.file.File;
+import com.hedera.node.app.service.file.impl.codec.FileServiceStateTranslator;
+import com.hedera.node.app.service.file.impl.codec.FileServiceStateTranslator.FileMetadataAndContent;
+import com.hedera.node.app.service.file.impl.test.FileTestBase;
+import com.hedera.node.app.service.mono.files.HFileMeta;
+import com.hedera.node.app.service.mono.utils.MiscUtils;
+import org.apache.commons.codec.DecoderException;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
+
+@ExtendWith(MockitoExtension.class)
+public class FileServiceStateTranslatorTest extends FileTestBase {
+
+    @Mock
+    private com.hedera.node.app.service.mono.files.HederaFs hederaFs;
+
+    @BeforeEach
+    void setUp() {}
+
+    @Test
+    void createFileMetadataAndContentFromFile() throws DecoderException {
+        final var existingFile = readableStore.getFileMetadata(fileId);
+        assertFalse(existingFile.deleted());
+
+        final FileMetadataAndContent convertedFile = FileServiceStateTranslator.pbjToState(file);
+
+        assertArrayEquals(
+                convertedFile.data(), getExpectedMonoFileMetaAndContent().data());
+        assertEquals(
+                convertedFile.metadata().getExpiry(),
+                getExpectedMonoFileMetaAndContent().metadata().getExpiry());
+        assertEquals(
+                convertedFile.metadata().getMemo(),
+                getExpectedMonoFileMetaAndContent().metadata().getMemo());
+        assertEquals(
+                MiscUtils.describe(convertedFile.metadata().getWacl()),
+                MiscUtils.describe(
+                        getExpectedMonoFileMetaAndContent().metadata().getWacl()));
+        assertEquals(
+                convertedFile.metadata().isDeleted(),
+                getExpectedMonoFileMetaAndContent().metadata().isDeleted());
+    }
+
+    @Test
+    void createFileMetadataAndContentFromReadableFileStore() throws DecoderException {
+        final var existingFile = readableStore.getFileMetadata(fileId);
+        assertFalse(existingFile.deleted());
+
+        final FileMetadataAndContent convertedFile = FileServiceStateTranslator.pbjToState(fileId, readableStore);
+
+        assertArrayEquals(
+                convertedFile.data(), getExpectedMonoFileMetaAndContent().data());
+        assertEquals(
+                convertedFile.metadata().getExpiry(),
+                getExpectedMonoFileMetaAndContent().metadata().getExpiry());
+        assertEquals(
+                convertedFile.metadata().getMemo(),
+                getExpectedMonoFileMetaAndContent().metadata().getMemo());
+        assertEquals(
+                MiscUtils.describe(convertedFile.metadata().getWacl()),
+                MiscUtils.describe(
+                        getExpectedMonoFileMetaAndContent().metadata().getWacl()));
+        assertEquals(
+                convertedFile.metadata().isDeleted(),
+                getExpectedMonoFileMetaAndContent().metadata().isDeleted());
+    }
+
+    @Test
+    void createFileFromMetadataContentAndFileId() {
+        final byte[] data = contents;
+        final com.hedera.node.app.service.mono.files.HFileMeta metadata =
+                new HFileMeta(true, jKeyList, expirationTime, memo);
+
+        final com.hederahashgraph.api.proto.java.FileID fileID = monoFileID;
+
+        final File convertedFile = FileServiceStateTranslator.stateToPbj(data, metadata, fileID);
+
+        assertEquals(createFile(), convertedFile);
+    }
+
+    @Test
+    void createFileFromFileIDAndHederaFs() {
+
+        final com.hederahashgraph.api.proto.java.FileID fileID = monoFileID;
+        final com.hedera.node.app.service.mono.files.HFileMeta metadata =
+                new HFileMeta(true, jKeyList, expirationTime, memo);
+
+        given(hederaFs.cat(fileID)).willReturn(contents);
+        given(hederaFs.getattr(fileID)).willReturn(metadata);
+        final File convertedFile = FileServiceStateTranslator.stateToPbj(fileID, hederaFs);
+
+        assertEquals(createFile(), convertedFile);
+    }
+
+    private FileMetadataAndContent getExpectedMonoFileMetaAndContent() throws DecoderException {
+        var keys = com.hedera.node.app.service.mono.legacy.core.jproto.JKey.convertKey(
+                Key.newBuilder().keyList(file.keys()).build(), 1);
+        com.hedera.node.app.service.mono.files.HFileMeta hFileMeta =
+                new HFileMeta(file.deleted(), keys, file.expirationTime(), file.memo());
+        return new FileMetadataAndContent(file.contents().toByteArray(), hFileMeta);
+    }
+}
